requirements.txt
# HYDRA Trading Bot — Python Dependencies

# Core runtime
loguru>=0.7.0
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.1.0
requests>=2.31.0

# Dashboard (FastAPI + Uvicorn — dashboard.py uses FastAPI, not Dash)
fastapi>=0.110.0
uvicorn>=0.29.0

# Exchange / market data
ccxt>=4.1.0
yfinance>=0.2.36

# Telegram
python-telegram-bot>=20.0

# ML (optional — set ml.enabled: true in config.yaml to activate)
xgboost>=2.0.0
scikit-learn>=1.4.0

# Utilities
python-dateutil>=2.8.0
